+++ src/main/java/com/airbus_cyber_security/graylog/wizard/alert/rest/models/responses/GetDataAlertRule.java
@@ -30,6 +30,7 @@ import javax.validation.constraints.NotNull;
 import java.util.Map;
 
 // TODO either inherit or reuse AlertRuleRequest to avoid duplication in this class => I believe it should be the same object, with optional fields maybe...
+// TODO instead of allowing some Nullable fields, such as condition, stream, condition_parameters, maybe should be either corrupted or contain correct information
 @AutoValue
 @JsonAutoDetect
 public abstract class GetDataAlertRule {